+++ src/components/menu/__internal__/submenu/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./submenu.component";
+export type { SubmenuProps } from "./submenu.component";