+++ src/components/menu/__internal__/submenu/submenu.context.js
@@ -1,4 +1,0 @@
-import React from "react";
-
-const SubmenuContext = React.createContext({});
-export default SubmenuContext;