+++ src/components/menu/index.js
@@ -1,6 +1,0 @@
-export { default as Menu } from "./menu.component";
-export { default as MenuItem } from "./menu-item/menu-item.component";
-export { default as MenuDivider } from "./menu-divider/menu-divider.component";
-export { default as MenuSegmentTitle } from "./menu-segment-title/menu-segment-title.component";
-export { default as MenuFullscreen } from "./menu-full-screen/menu-full-screen.component";
-export { default as ScrollableBlock } from "./scrollable-block/scrollable-block.component";