+++ src/components/menu/menu-divider/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./menu-divider";
-export * from "./menu-divider";