+++ src/components/menu/menu-item/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./menu-item";
-export * from "./menu-item";