+++ src/components/menu/menu-segment-title/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./menu-segment-title";
-export * from "./menu-segment-title";