+++ src/components/menu/menu.context.js
@@ -1,8 +1,0 @@
-import React from "react";
-
-export default React.createContext({
-  menuType: "light",
-  inMenu: false,
-  openSubmenuId: null,
-  setOpenSubmenuId: /* istanbul ignore next */ () => {},
-});