+++ src/components/message/message-content/message-content.style.ts
@@ -1,7 +1,11 @@
 import styled from "styled-components";
+import { MessageContentProps } from "./message-content.component";
 
-const MessageContentStyle = styled.div`
-  padding: 15px 50px 15px 20px;
+const MessageContentStyle = styled.div<
+  Pick<MessageContentProps, "showCloseIcon">
+>`
+  padding: 15px ${({ showCloseIcon }) => (showCloseIcon ? "50px" : "20px")} 15px
+    20px;
   white-space: pre-wrap;
   flex: 1;